--- shopping-cart-java-node1/shopping-cart/src/main/java/com/example/shoppingcart/typedimpl/ShoppingCartException.java
@@ -1,20 +1,0 @@
-package com.example.shoppingcart.typedimpl;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.lightbend.lagom.serialization.Jsonable;
-import lombok.EqualsAndHashCode;
-import lombok.Value;
-
-@Value
-@EqualsAndHashCode(callSuper = false)
-@JsonDeserialize
-public class ShoppingCartException extends RuntimeException implements Jsonable {
-    public final String message;
-
-    @JsonCreator
-    public ShoppingCartException(String message) {
-        super(message);
-        this.message = message;
-    }
-}
