--- shopping-cart-java-node1/shopping-cart/src/main/java/com/example/shoppingcart/typedimpl/ShoppingCartModule.java
@@ -1,16 +1,0 @@
-package com.example.shoppingcart.typedimpl;
-
-import com.example.shoppingcart.api.ShoppingCartService;
-import com.google.inject.AbstractModule;
-import com.lightbend.lagom.javadsl.server.ServiceGuiceSupport;
-
-/**
- * The module that binds the {@link ShoppingCartService} so that it can be served.
- */
-public class ShoppingCartModule extends AbstractModule implements ServiceGuiceSupport {
-    @Override
-    protected void configure() {
-        bindService(ShoppingCartService.class, ShoppingCartServiceImpl.class);
-        bind(ReportRepository.class);
-    }
-}
