+++ shopping-cart-java-node1/shopping-cart/src/main/resources/application.conf
@@ -61,3 +61,7 @@ akka {
     # auto-down-unreachable-after = 10s
   }
 }
+
+# FIXME needed for first join to node without akka-cluster-typed
+# This should be fixed in Akka
+akka.cluster.configuration-compatibility-check.enforce-on-join = off
